% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-AA_sequence}
\alias{data-AA_sequence}
\alias{AA_sequence}
\title{Data set containing multiple sequence alignment information}
\format{
AAbin
}
\description{
This data set contains multiple sequence alignment information  to be
attached to the tree.\cr
Chap.7.4
}
\examples{
data(AA_sequence)
}
\keyword{data}
