% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-alleles.csv}
\alias{data-alleles.csv}
\alias{alleles.csv}
\title{Allele table}
\format{
A data frame contains 386 rows and 385 variables, with first row of
tips labels.
}
\source{
\url{https://raw.githubusercontent.com/katholt/plotTree/master/
tree_example_april2015/alleles.csv}
}
\description{
The allele table with original data to be processed to SNP data. Then attached
to the tree after visualizing.\cr
Chap.7.2
}
\examples{
data(alleles.csv)
}
\keyword{data}
