% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-alltax_info.csv}
\alias{data-alltax_info.csv}
\alias{alltax_info.csv}
\title{Data frame containing hierarchical relationship}
\format{
A data frame contains 1351 rows and 7 variables.
}
\source{
\url{ https://doi.org/10.1101/gr.126573.111}
}
\description{
This Data frame contains hierarchical relationship to be visualized with
another data frame that contains taxa and factor information and pvalue,
using \code{ggdiffcalde()} of The \code{MicrobiotaProcess} package.\cr
It's the analysis result of \code{diff_analysis()} using public datasets
(Kostic et al. 2012) with The \code{MicrobiotaProcess} package.\cr
These data frame are used to demonstrates how to visualize the differential
taxonomy tree with The \code{MicrobiotaProcess} package.\cr
Chap.11.1
}
\examples{
data(alltax_info.csv)
}
\keyword{data}
