% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-anole.tre}
\alias{data-anole.tre}
\alias{anole.tre}
\title{Tree to color with continuous state transition in edges}
\format{
S3 class:phylo
}
\source{
\url{http://www.phytools.org/eqg2015/data/anole.tre}
}
\description{
This tree is used to demonstrate the functionality of plotting continuous
state transition in edges using the \code{continuous} parameter of \code{ggtree}.\cr
Chap.4.3.5
}
\examples{
data(anole.tre)
}
\keyword{data}
