% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-barplot_attr.csv}
\alias{data-barplot_attr.csv}
\alias{barplot_attr.csv}
\title{Data set of the abundance of microbes at the body sites of greatest
prevalence}
\format{
A data frame contains 332 rows and 3 variables, with a column of
"node" or first column of taxa labels.
}
\source{
\url{https://doi.org/10.7717/peerj.1029}
}
\description{
The data set of abundance of microbes at the body sites of greatest
prevalence to be visualized using barplot and attached to the tree.\cr
Chap.10.3
}
\examples{
data(barplot_attr.csv)
}
\keyword{data}
