% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_barplot_attr}
\alias{data-df_barplot_attr}
\alias{df_barplot_attr}
\title{Data set of the abundance of microbes at the body sites of greatest
prevalence}
\format{
A data frame contains 332 rows and 3 variables, with a column of
"node" or first column of taxa labels.
}
\description{
The data set of abundance of microbes at the body sites of greatest
prevalence to be visualized using barplot and attached to the tree.\cr
Chap.10.3
}
\examples{
data(df_barplot_attr)
}
\references{
Asnicar, Francesco, George Weingart, Timothy L Tickle, Curtis Huttenhower, and Nicola Segata.
Compact Graphical Representation of Phylogenetic Data and Metadata with GraPhlAn.
PeerJ, 3: e1029, June 2015.
\doi{10.7717/peerj.1029}
}
\keyword{data}
