% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_difftax}
\alias{data-df_difftax}
\alias{df_difftax}
\title{Data frame containing taxa and factor information and pvalue}
\format{
A data frame contains 36 rows and 3 variables.
}
\description{
This Data frame contains taxa and factor information and pvalue to be visualized
with another data frame that contains hierarchical relationship, using
\code{ggdiffcalde()} of The \code{MicrobiotaProcess} package.\cr
These data frame are used to demonstrates how to visualize the differential
taxonomy tree with The \code{MicrobiotaProcess} package.\cr
Chap.11.1
}
\examples{
data(df_difftax)
}
\references{
Kostic, Aleksandar D, Dirk Gevers, Chandra Sekhar Pedamallu, Monia Michaud, Fujiko Duke, Ashlee M Earl, Akinyemi I Ojesina, et al.
Genomic Analysis Identifies Association of Fusobacterium with Colorectal Carcinoma.”
Genome Research, Volume 22, Issue 2, 2012, Pages 292–98.
\doi{10.1101/gr.126573.111}
}
\keyword{data}
