% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-df_inode_data}
\alias{data-df_inode_data}
\alias{df_inode_data}
\title{Nodedata to be mapped to tree}
\format{
A data frame with 6 rows and 6 variables with a column of node
or first column of taxa labels.
}
\source{
\url{https://raw.githubusercontent.com/TreeViz/metastyle/master/design/
viz_targets_exercise/inode_data.csv}
}
\description{
This data contains annotation information about nodes of "tree_boots.nwk",
that is to be mapped to the tree using the \verb{\%<+\%} operator.\cr
Chap.7.1
}
\examples{
data(df_inode_data)
}
\keyword{data}
