% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_NJIDqgsS}
\alias{data-df_NJIDqgsS}
\alias{df_NJIDqgsS}
\title{Population genetics data No.2}
\format{
A data frame contains 1351 rows and 7 variables, with a column of
"node" or first column of taxa labels.
}
\description{
Population genetics data to be visualized using multiple graphs and
attached to tree No.2 with \code{geo_fruit()} of \code{ggtreeExtra}.\cr
Chap.10.4
}
\examples{
data(df_NJIDqgsS)
}
\references{
Wong, Vanessa K., Stephen Baker, Derek J. Pickard, Julian Parkhill, Andrew J. Page, Nicholas A. Feasey, Robert A. Kingsley, et al.
Phylogeographical Analysis of the Dominant Multidrug-Resistant H58 Clade of Salmonella Typhi Identifies Inter- and Intracontinental Transmission Events.
Nature Genetics, Volume 47, Issue 6, May 2015, Pages 632–39.
\doi{10.1038/ng.3281}
}
\keyword{data}
