% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_ring_heatmap}
\alias{data-df_ring_heatmap}
\alias{df_ring_heatmap}
\title{Data set of the abundance and types of microbes}
\format{
A data frame contains 2324 rows and 3 variables, with a column of
"node" or first column of taxa labels.
}
\description{
The data set of abundance of microbes at different body sites to be
visualized using heatmap and attached to the tree.\cr
Chap.10.3
}
\examples{
data(df_ring_heatmap)
}
\references{
Asnicar, Francesco, George Weingart, Timothy L Tickle, Curtis Huttenhower, and Nicola Segata.
Compact Graphical Representation of Phylogenetic Data and Metadata with GraPhlAn.
PeerJ, 3: e1029, June 2015.
\doi{10.7717/peerj.1029}
}
\keyword{data}
