% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_svl}
\alias{data-df_svl}
\alias{df_svl}
\title{Data frame containing slv information}
\format{
A data frame contains 100 rows and 1 variables.
}
\source{
\url{http://www.phytools.org/eqg2015/data/svl.csv}
}
\description{
Data frame containing slv information to scale the color of the tree using
the \code{continuous} parameter of \code{ggtree} to plot continuous state transition
in edges.\cr
Chap.4.3.5. This file is provided in \url{http://www.phytools.org/eqg2015/asr.html}.
}
\examples{
data(df_svl)
}
\keyword{data}
