% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_tip_data}
\alias{data-df_tip_data}
\alias{df_tip_data}
\title{Tipdata to be mapped to tree}
\format{
A data frame contains 7 rows and 9 variables, with a column of
"node" or first column of taxa labels.
}
\source{
\url{https://raw.githubusercontent.com/TreeViz/metastyle/master/design/
viz_targets_exercise/tip_data.csv}
}
\description{
This data contains annotation information about tips of \code{tree_boots},
that is to be mapped to the tree using the \verb{\%<+\%} operator.\cr
Chap.7.1
}
\examples{
data(df_tip_data)
}
\keyword{data}
