% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_tippoint}
\alias{data-df_tippoint}
\alias{df_tippoint}
\title{Data set of the abundance and types of microbes}
\format{
A data frame contains 332 rows and 4 variables,with a column of
"node" or first column of taxa labels.
}
\description{
The data set of abundance and types of microbes to be attached to the tree
tips, provided by GraPhlAn (Asnicar et al. 2015).\cr
Chap.10.3
}
\examples{
data(df_tippoint)
}
\references{
Asnicar, Francesco, George Weingart, Timothy L Tickle, Curtis Huttenhower, and Nicola Segata.
Compact Graphical Representation of Phylogenetic Data and Metadata with GraPhlAn.
PeerJ, 3: e1029, June 2015.
\doi{10.7717/peerj.1029}
}
\keyword{data}
