% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-difftax_tab.csv}
\alias{data-difftax_tab.csv}
\alias{difftax_tab.csv}
\title{Data frame containing taxa and factor information and pvalue}
\format{
A data frame contains 36 rows and 3 variables.
}
\source{
\url{ https://doi.org/10.1101/gr.126573.111}
}
\description{
This Data frame contains taxa and factor information and pvalue to be visualized
with another data frame that contains hierarchical relationship, using
\code{ggdiffcalde()} of The \code{MicrobiotaProcess} package.\cr
These data frame are used to demonstrates how to visualize the differential
taxonomy tree with The \code{MicrobiotaProcess} package.\cr
Chap.11.1
}
\examples{
data(difftax_tab.csv)
}
\keyword{data}
