% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-dna_HPV58_aln}
\alias{data-dna_HPV58_aln}
\alias{dna_HPV58_aln}
\title{A DNAStringSet class to store the aligned sequnces of species of HPV58.tree}
\format{
S4 class:DNAStringSet
}
\source{
\url{https://doi.org/10.1128/JVI.01285-17}
}
\description{
A DNAStringSet class to store the aligned sequnces of species of HPV58.tree
}
\examples{
data(dna_HPV58_aln)
}
\keyword{data}
