% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-dna_sequence}
\alias{data-dna_sequence}
\alias{dna_sequence}
\title{Data set containing multiple sequence alignment information}
\format{
DNAbin
}
\description{
This data set contains multiple sequence alignment information  to be
attached to the tree.\cr
Chap.7.4
}
\examples{
data(dna_sequence)
}
\keyword{data}
