% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-gg_ladderize_example}
\alias{data-gg_ladderize_example}
\alias{gg_ladderize_example}
\title{Data set containing ladderize and nonladderize trees}
\format{
S3:gg, ggplot
}
\description{
This data set contains multiple graphs that demonstrate how to use
\code{plot.phylo()} and \code{ggtree} to visualize ladderize and nonladderize trees.\cr
Appendix.A.5.1
}
\examples{
data(gg_ladderize_example)
}
\keyword{data}
