% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-hmptree.nwk}
\alias{data-hmptree.nwk}
\alias{hmptree.nwk}
\title{Tree to display with multiple graphs for multi-dimensional data}
\format{
S3 class:phylo
}
\source{
\url{https://doi.org/10.1016/J.TIG.2012.09.005}
}
\description{
The tree to display with multiple graphs, for demonstrating the feature of
adding multiple layers to present different types of data by \code{ggtreeExtra}.\cr
Chap.10.3
}
\examples{
data(hmptree.nwk)
}
\keyword{data}
