% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-HPV58.tree}
\alias{data-HPV58.tree}
\alias{HPV58.tree}
\title{Tree to display with dot and line plots of pairwise nucleotide sequence distances}
\format{
S3 class:phylo
}
\source{
\url{https://doi.org/10.1128/JVI.01285-17}
}
\description{
This tree to provide accession numbers from its tip labels, which are used to
calculates pairwise nucleotide sequence distances. The distance matrix
then is visualized as dot and line plots and attached to the tree,
demonstrating the abilities of \code{facet_plot()} of ggtree that can add
multiple layers to a specific panel.\cr
Reproduction fo Fig. 1 of (Chen et al. 2017).\cr
Chap.13.1
}
\examples{
data(HPV58.tree)
}
\keyword{data}
