% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-info.csv}
\alias{data-info.csv}
\alias{info.csv}
\title{Sampling information data set}
\format{
A data frame contains 386 rows and 6 variables, with a column of
"node" or first column of taxa labels.
}
\source{
\url{https://raw.githubusercontent.com/katholt/plotTree/master/
tree_example_april2015/info.csv}
}
\description{
The sampling information data set be attached to the tree tip. \cr
Chap.7.2
}
\examples{
data(info.csv)
}
\keyword{data}
