% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-NJIDqgsS_data.csv}
\alias{data-NJIDqgsS_data.csv}
\alias{NJIDqgsS_data.csv}
\title{Population genetics data No.2}
\format{
A data frame contains 1351 rows and 7 variables, with a column of
"node" or first column of taxa labels.
}
\source{
\url{https://doi.org/10.1038/ng.3281}
}
\description{
Population genetics data to be visualized using multiple graphs and
attached to tree No.2 with \code{geo_fruit()} of \code{ggtreeExtra}.\cr
Chap.10.4
}
\examples{
data(NJIDqgsS_data.csv)
}
\keyword{data}
