% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-NJIDqgsS_tree.nwk}
\alias{data-NJIDqgsS_tree.nwk}
\alias{NJIDqgsS_tree.nwk}
\title{Tree No.2 to display with visualized population genetics data}
\format{
S3 class:phylo
}
\source{
\url{https://doi.org/10.1038/ng.3281}
}
\description{
Tree No.2 to display with visualized population genetics data for
demonstrating the utility of \code{ggtreeExtra}.\cr
Reproduction of Fig 1 of (Wong et al. 2015).\cr
Chap.10.4
}
\examples{
data(NJIDqgsS_tree.nwk)
}
\keyword{data}
