% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-sequence.fasta}
\alias{data-sequence.fasta}
\alias{sequence.fasta}
\title{Data set containing multiple sequence alignment information}
\format{
DNAbin
}
\description{
This data set contains multiple sequence alignment information  to be
attached to the tree.\cr
Chap.7.4
}
\examples{
data(sequence.fasta)
}
\keyword{data}
