% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-svl.csv}
\alias{data-svl.csv}
\alias{svl.csv}
\title{Data frame containing slv information}
\format{
A data frame contains 100 rows and 1 variables.
}
\source{
\url{http://www.phytools.org/eqg2015/data/anole.tre}
}
\description{
Data frame containing slv information to scale the color of the tree using
the \code{continuous} parameter of \code{ggtree} to plot continuous state transition
in edges.\cr
Chap.4.3.5
}
\examples{
data(svl.csv)
}
\keyword{data}
