% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-test.nw}
\alias{data-test.nw}
\alias{test.nw}
\title{Tree to reroot}
\format{
S4 class:treedata
}
\description{
This tree is used to demonstrate the \code{reroot()} method for treedata object
to re-root a phylogenetic tree with associated data.\cr
Chap.2.3
}
\examples{
data(test.nw)
}
\keyword{data}
