% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-text_RMI_tree}
\alias{data-text_RMI_tree}
\alias{text_RMI_tree}
\title{Tree to display with symbolic points indicating partitioned bootstrap values}
\format{
character
}
\description{
This tree is to display with symbolic points with different color or shape
indicating the bootstrap values belong to which category.\cr
Catagories are defined by cutting the bootstrap values into several interval
to indicate whether the clade is high, moderate or low support.\cr
Chap.13.2
}
\examples{
data(text_RMI_tree)
}
\keyword{data}
