% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-tippoint_attr.csv}
\alias{data-tippoint_attr.csv}
\alias{tippoint_attr.csv}
\title{Data set of the abundance and types of microbes}
\format{
A data frame contains 332 rows and 4 variables,with a column of
"node" or first column of taxa labels.
}
\source{
\url{https://doi.org/10.7717/peerj.1029}
}
\description{
The data set of abundance and types of microbes to be attached to the tree
tips.\cr
Chap.10.3
}
\examples{
data(tippoint_attr.csv)
}
\keyword{data}
