% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-Tree_30.4.19.nwk}
\alias{data-Tree_30.4.19.nwk}
\alias{Tree_30.4.19.nwk}
\title{Tree to group and highlight}
\format{
S3 class:Phylo
}
\source{
\url{https://doi.org/10.1016/j.molimm.2019.07.022.}
}
\description{
This tree is used to reproduce Figure 1 of (Larsen et al. 2019) with
grouping, highlighting and labelling.\cr
Chap.13.3
}
\examples{
data(Tree_30.4.19.nwk)
}
\keyword{data}
