% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_anole}
\alias{data-tree_anole}
\alias{tree_anole}
\title{Tree to color with continuous state transition in edges}
\format{
S3 class:phylo
}
\source{
\url{http://www.phytools.org/eqg2015/data/anole.tre}
}
\description{
This tree is used to demonstrate the functionality of plotting continuous
state transition in edges using the \code{continuous} parameter of \code{ggtree}.\cr
Chap.4.3.5. This file is provided in \url{http://www.phytools.org/eqg2015/asr.html}.
}
\examples{
data(tree_anole)
}
\keyword{data}
