% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_boots}
\alias{data-tree_boots}
\alias{tree_boots}
\title{Tree to be mapped to data}
\format{
S3 class:phylo
}
\source{
\url{https://raw.githubusercontent.com/TreeViz/metastyle/master/design/
viz_targets_exercise/tree_boots.nwk}
}
\description{
This tree is used to demonstrate the functionality of the \verb{\%<+\%} operator
to attach annotation data to a ggtree graphic object.\cr
Chap.7.1
}
\examples{
data(tree_boots)
}
\keyword{data}
