% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_Candidaauris}
\alias{data-tree_Candidaauris}
\alias{tree_Candidaauris}
\title{Tree No.1 to display with visualized population genetics data}
\format{
S3 class:phylo
}
\source{
\url{https://doi.org/10.1128/mBio.03364-19}
}
\description{
Tree No.1 to display with visualized population genetics data for
demonstrating the utility of \code{ggtreeExtra}.\cr
Reproduction of Fig.4 of (Chow et al. 2020).\cr
Chap.10.4
}
\examples{
data(tree_Candidaauris)
}
\keyword{data}
