% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_HPV58}
\alias{data-tree_HPV58}
\alias{tree_HPV58}
\title{Tree to display with dot and line plots of pairwise nucleotide sequence distances}
\format{
S3 class:phylo
}
\description{
This tree is to provide accession numbers from its tip labels, which are used to
calculates pairwise nucleotide sequence distances. The distance matrix
then is visualized as dot and line plots and attached to the tree,
demonstrating the abilities of \code{facet_plot()} of ggtree that can add
multiple layers to a specific panel.\cr
Reproduction fo Fig. 1 of (Chen et al. 2017).\cr
Chap.13.1
}
\examples{
data(tree_HPV58)
}
\references{
Chen, Zigui, Wendy C. S. Ho, Siaw Shi Boon, Priscilla T. Y. Law, Martin C. W. Chan, Rob DeSalle, Robert D. Burk, and Paul K. S. Chan.
Ancient Evolution and Dispersion of Human Papillomavirus 58 Variants.
Journal of Virology, Volume 91, Issue 21, November 2017, e01285–17.
\doi{10.1128/JVI.01285-17}
}
\keyword{data}
