% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_long_branch_example}
\alias{data-tree_long_branch_example}
\alias{tree_long_branch_example}
\title{Tree to display with shrunk outlier long branch}
\format{
S3 class:Phylo
}
\description{
This tree is to display with display with shrunk outlier long branch to
demonstrate how to shrink outlier long branch with \code{ggtree} by modifying
coordination of the out groups, or with the \code{ggbreak} package.\cr
Appendix.A.5.3
}
\examples{
data(tree_long_branch_example)
}
\keyword{data}
