% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_nex}
\alias{data-tree_nex}
\alias{tree_nex}
\title{Tree to display with silhouettes from Phylopic}
\format{
S3 class:Phylo
}
\description{
This tree is to display with display with silhouettes from Phylopic,
demonstrating the feature of \code{ggtree} that will automatically download
phylopic figures by querying provided UID.\cr
Chap.8.2
}
\examples{
data(tree_nex)
}
\keyword{data}
