% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_NJIDqgsS}
\alias{data-tree_NJIDqgsS}
\alias{tree_NJIDqgsS}
\title{Tree No.2 to display with visualized population genetics data}
\format{
S3 class:phylo
}
\description{
Tree No.2 to display with visualized population genetics data for
demonstrating the utility of \code{ggtreeExtra}.\cr
Reproduction of Fig 1 of (Wong et al. 2015).\cr
Chap.10.4
}
\examples{
data(tree_NJIDqgsS)
}
\references{
Wong, Vanessa K., Stephen Baker, Derek J. Pickard, Julian Parkhill, Andrew J. Page, Nicholas A. Feasey, Robert A. Kingsley, et al.
Phylogeographical Analysis of the Dominant Multidrug-Resistant H58 Clade of Salmonella Typhi Identifies Inter- and Intracontinental Transmission Events.
Nature Genetics, Volume 47, Issue 6, May 2015, Pages 632–39.
\doi{10.1038/ng.3281}
}
\keyword{data}
