% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_nodes.newick}
\alias{data-tree_nodes.newick}
\alias{tree_nodes.newick}
\title{Tree to reroot}
\format{
S4 class:treedata
}
\source{
\url{https://github.com/stamatak/tree-viz-issues}
}
\description{
This tree is used to demonstrate the \code{root()} method for treedata object
to re-root a phylogenetic tree with associated data.\cr
Chap.2.3
}
\examples{
data(tree_nodes.newick)
}
\references{
Lucas Czech, Jaime Huerta-Cepas, Alexandros Stamatakis,
A Critical Review on the Use of Support Values in Tree Viewers and Bioinformatics Toolkits.
Molecular Biology and Evolution, Volume 34, Issue 6, June 2017, Pages 1535–1542,
https://doi.org/10.1093/molbev/msx055
}
\keyword{data}
