% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_nwk}
\alias{data-tree_nwk}
\alias{tree_nwk}
\title{Tree to display with sampling information, SNP and Trait data}
\format{
S3 class:phylo
}
\source{
\url{https://raw.githubusercontent.com/katholt/plotTree/master/
tree_example_april2015/tree.nwk}
}
\description{
The tree to display with sampling information, attached with dot chart of SNP
and bar chart of Trait data.\cr
Chap.7.2
}
\examples{
data(tree_nwk)
}
\keyword{data}
