% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_seq_nwk}
\alias{data-tree_seq_nwk}
\alias{tree_seq_nwk}
\title{Tree to display with multiple sequence alignment}
\format{
S3 class:phylo
}
\description{
This tree is used to display with multiple sequence alignment.\cr
Chap.7.4
}
\examples{
data(tree_seq_nwk)
}
\keyword{data}
