% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDbook-package.R
\docType{package}
\name{TDbook-package}
\alias{TDbook}
\alias{TDbook-package}
\title{TDbook: Companion Package for the Book "Data Integration, Manipulation and Visualization of Phylogenetic Trees" by Guangchuang Yu (2022, ISBN:9781032233574)}
\description{
The companion package that provides all the datasets used in the book "Data Integration, Manipulation and Visualization of Phylogenetic Trees" by Guangchuang Yu (2022, ISBN:9781032233574).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/TDBook}
  \item Report bugs at \url{https://github.com/YuLab-SMU/TDbook/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})
  \item Lin Li \email{williamlee2220@qq.com} (\href{https://orcid.org/0000-0002-6522-1889}{ORCID})
}

}
\keyword{internal}
