% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\encoding{UTF-8}
\name{data-df_alltax_info}
\alias{data-df_alltax_info}
\alias{df_alltax_info}
\title{Data frame containing hierarchical relationship}
\format{
A data frame to store taxonomy information, contains 1351 rows and 7
variables (classification levels).
}
\description{
This Data frame contains hierarchical relationship to be visualized with
another data frame that contains taxa and factor information and pvalue,
using \code{ggdiffcalde()} of The \code{MicrobiotaProcess} package.\cr
It's the analysis result of \code{diff_analysis()} using public datasets
(Kostic et al. 2012) with The \code{MicrobiotaProcess} package.\cr
These data frame are used to demonstrates how to visualize the differential
taxonomy tree with The \code{MicrobiotaProcess} package.\cr
Chap.11.1
}
\examples{
data(df_alltax_info)
}
\references{
Kostic, Aleksandar D, Dirk Gevers, Chandra Sekhar Pedamallu, Monia Michaud, Fujiko Duke, Ashlee M Earl, Akinyemi I Ojesina, et al.
Genomic Analysis Identifies Association of Fusobacterium with Colorectal Carcinoma.”
Genome Research, Volume 22, Issue 2, 2012, Pages 292–98.
\doi{10.1101/gr.126573.111}
}
\keyword{data}
