% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-dna_HPV58_aln}
\alias{data-dna_HPV58_aln}
\alias{dna_HPV58_aln}
\title{A DNAbin class to store the aligned sequnces of species of HPV58.tree}
\format{
S3 class:DNAbin
}
\description{
A DNAbin class to store the aligned sequnces of species of HPV58.tree
}
\examples{
data(dna_HPV58_aln)
}
\references{
Chen, Zigui, Wendy C. S. Ho, Siaw Shi Boon, Priscilla T. Y. Law, Martin C. W. Chan, Rob DeSalle, Robert D. Burk, and Paul K. S. Chan.
Ancient Evolution and Dispersion of Human Papillomavirus 58 Variants.
Journal of Virology, Volume 91, Issue 21, November 2017, e01285–17.
\doi{10.1128/JVI.01285-17}
}
\keyword{data}
