% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree.nwk}
\alias{data-tree.nwk}
\alias{tree.nwk}
\title{tree (phylo class) to display with multiple sequence alignment}
\description{
This tree is used to display with multiple sequence alignment
}
\examples{
data(tree.nwk)
}
\keyword{data}
