% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_Candidaauris}
\alias{data-tree_Candidaauris}
\alias{tree_Candidaauris}
\title{Tree No.1 to display with visualized population genetics data}
\format{
S3 class:phylo
}
\description{
Tree No.1 to display with visualized population genetics data for
demonstrating the utility of \code{ggtreeExtra}.\cr
Reproduction of Fig.4 of (Chow et al. 2020).\cr
Chap.10.4
}
\examples{
data(tree_Candidaauris)
}
\references{
Chow, Nancy A., José F. Muñoz, Lalitha Gade, Elizabeth L. Berkow, Xiao Li, Rory M. Welsh, Kaitlin Forsberg, et al.
Tracing the Evolutionary History and Global Expansion of Candida Auris Using Population Genomic Analyses.
Edited by Geraldine Butler and Kirsten Nielsen.
mBio, Volume 11, Issue 2, April 2020.
\doi{10.1128/mBio.03364-19}
}
\keyword{data}
