% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllData.R
\docType{data}
\name{data-tree_hmptree}
\alias{data-tree_hmptree}
\alias{tree_hmptree}
\title{Tree to display with multiple graphs for multi-dimensional data}
\format{
S3 class:phylo
}
\description{
This tree is to display with multiple graphs, and by reproducing
Fig.2 of (Morgan, Segata, and Huttenhower 2013), demonstrates the feature of
adding multiple layers to present different types of data by \code{ggtreeExtra}.\cr
Chap.10.3
}
\examples{
data(tree_hmptree)
}
\references{
Morgan, Xochitl C., Nicola Segata, and Curtis Huttenhower.
Biodiversity and Functional Genomics in the Human Microbiome.
Trends in Genetics Volume 29, Issue 1, 2013 January, Pages 51–58.
\doi{10.1016/J.TIG.2012.09.005}
}
\keyword{data}
